GDB-5337 Fix getting current repo from local storage
Untill recently the currently selected repository was stored as a cookie by GraphDB workbench . Now it is stored in local storage.
Refactored the CookiesService to LocalStorageService.
Added local storage keys as constants.
Removed current repo port part as with local storage it is not needed and is removed from GraphDB workbench.