OR-100 Updates the handler for the SPARQL button
- Updated the handler for the SPARQL button to correspond to the changes
in the backend. This is the small part of the changes that will be done
to the Mapping UI due to the splitting of OntoRefine and GraphDB.

When the GraphDB is secured, the users will be redirected to the
Workbench login form, if there isn't active session. Later we may change
this behavior with something more user friendly.