Fixes the setting of the RDF type for blank nodes
- The issue was related to the initialization of the objects in the
model, when the type is added. The array which should contain the type
mapping was actually detached from the actual model so no update was
detected when the events were published. The disappearance in the view
was also related to that issue, because the view is refreshed after the
update event is published.
- Added tests to verify the correct behavior when adding RDF type to the
blank node and also to prevent regressions in the future.
- Added few minor fixes in the code documentation. By convention all
sentences should end with full stop character.
- Updated `.gitignore` file to exclude the files from the `.angular`
directory, which is generated when we run the cypress tests locally.
- Fixed flaky test.