Introduces the SPARQL Query Editor
- Added new component `sparql-editor` based on the Ontotext's wrapper
implementation of Yasgui editor -
[ontotext-yasgui](https://github.com/Ontotext-AD/ontotext-yasgui).

  Now the view of the RDF Mapper is split into tabs, where the first
(initial) tab contains the `Visual RDF Mapper` and the second contains
the `SAPRQL Query Editor`.

  The editor allow the users to define an RDF mapping through SPARQL
`CONSTRUCT` queries directly in the Refine.

There were some improvements to the Refine virtual endpoint, which
allowed us to connect actual editor, which can execute various SPARQL
queries against the data stored for the project.

Previously this functionality was only available by connecting the
Refine to remote GraphDB instance, where the user had to open the
Workbench and create federated to the Refine endpoint.
The old functionality is kept unchanged, but now the new editor provides
a way to decouple Refine and GraphDB.

Editor functionalities:
1. Execution of SPARQL `SELECT`, `CONSTRUCT` and `ASK`(experimental)
queries.
2. Visualization of the results as tabular data.
3. Multiple query tabs.
4. Persistence of the state of the editor. This is achieved by storing
the editor configuration from the browser as part of the refine project
metadata. When the editor is open for a project, it will check if there
is a persisted configuration and use it, otherwise it will initialize
default editor with single tab. This functionality allows the users to
open and work on the same queries from different browsers/machines.
5. Generation of SPARQL templates. The editor provides a way to
generate SPARQL queries based on the columns in the refine project data
or to generate query based on the mapping defined in the `Visual RDF
Mapper`. The queries are opened in a new tab in the editor and the user
can start modifying them immediately.
6. Open the current query in the GraphDB, when a remote instance is
connected.
7. Options to download the currently active query or the active query
with the `SERVICE` clause included, as a file.
8. Download of the results of a `CONSTRUCT` query as a turtle (.ttl)
file.
9. And a lot more minor improvements and additional functionalities
related to the refine project and the data it contains.

  Overall we tried to theme it as close as possible to the GrapthDB
Workbench SPARQL editor for the functionalities that the Refine
supports. We kept the controls and functionalities for the editor as
close as possible to the `Virtual RDF Mapper` so that both tabs can give
the same feel while working on them.

- As addition we have discovered and fixed some issues in the RDF Mapper
as whole.

- Fixed some security vulnerabilities.