Fixes an issue related to Yasgui inability to work with relative URLs
- When the response from the execution of SPARQL query is too large for
Yasqe to render, it displays message in the area where the result table
should be.
  However, for some reason, while trying to display this message, the
logic tries to parse the provided URL for the SPARQL endpoint, which was
provided as relative path. That cause error to be thrown in the brower
console without any indication for it, which left the user with response
from the query that there are no results, which is inaccurate.

  So in order to fix the issue, the SPARQL endpoint was converted and
then provided as absolute URL to the Yasgui.